Move 90Test to existing one. And clean up Cache serialization
